test: use comprehensive testing for uncomparables
